requirements.txt
numpy==1.22.0
opencv_contrib_python==4.6.0.66
protobuf==4.22.0
Flask==2.2.3
Flask-Cors==3.0.10
requests==2.28.2
GitPython==3.1.31
opencv_python==4.6.0.66
Pillow==9.4.0